Update imports for the moved compat module
