+++ src/pip/_internal/baseparser.py
@@ -9,8 +9,8 @@ from distutils.util import strtobool
 
 from pip._vendor.six import string_types
 
-from pip._internal.compat import get_terminal_size
 from pip._internal.configuration import Configuration, ConfigurationError
+from pip._internal.utils.compat import get_terminal_size
 
 logger = logging.getLogger(__name__)
