+++ src/pip/_internal/commands/freeze.py
@@ -5,8 +5,8 @@ import sys
 from pip._internal import index
 from pip._internal.basecommand import Command
 from pip._internal.cache import WheelCache
-from pip._internal.compat import stdlib_pkgs
 from pip._internal.operations.freeze import freeze
+from pip._internal.utils.compat import stdlib_pkgs
 
 DEV_PKGS = {'pip', 'setuptools', 'distribute', 'wheel'}
