+++ src/pip/_internal/req/req_install.py
@@ -23,7 +23,6 @@ from pip._vendor.pkg_resources import RequirementParseError, parse_requirements
 
 from pip._internal import wheel
 from pip._internal.build_env import NoOpBuildEnvironment
-from pip._internal.compat import native_str
 from pip._internal.download import (
     is_archive_file, is_url, path_to_url, url_to_path,
 )
@@ -34,6 +33,7 @@ from pip._internal.locations import (
 from pip._internal.models.index import PyPI, TestPyPI
 from pip._internal.models.link import Link
 from pip._internal.req.req_uninstall import UninstallPathSet
+from pip._internal.utils.compat import native_str
 from pip._internal.utils.hashes import Hashes
 from pip._internal.utils.logging import indent_log
 from pip._internal.utils.misc import (
