+++ src/pip/_internal/utils/outdated.py
@@ -9,8 +9,8 @@ import sys
 from pip._vendor import lockfile, pkg_resources
 from pip._vendor.packaging import version as packaging_version
 
-from pip._internal.compat import WINDOWS
 from pip._internal.index import PackageFinder
+from pip._internal.utils.compat import WINDOWS
 from pip._internal.utils.filesystem import check_path_owner
 from pip._internal.utils.misc import ensure_dir, get_installed_version
